Add  serialization/deserialization to binary format. Add deserialization from byte[] array for C#. Fix deserialization problems for LSF job scheduler under windows.
work-items: #49193
git-tfs-id: [https://sede-pt-ce2.visualstudio.com/DefaultCollection]$/Basin Modeling/IBS/Trunk;C52403

+++ development/applications/casa/src/CmdAddVarPrm.C
@@ -345,7 +345,7 @@ void CmdAddVarPrm::execute( std::auto_ptr<casa::ScenarioAnalysis> & sa )
          }
          size_t numPts = atol( m_prms[pos++].c_str() ); // number of points for minimum multipoint perm. profile
 
-         minModelPrms.push_back( numPts );
+         minModelPrms.push_back( static_cast<double>( numPts ) );
          
          if ( pos + numPts * 2 >= m_prms.size() )
          {
